refactor: Move RestrictedControllerMessenger to separate file
The `RestrictedControllerMessenger` has been moved to a separate file.
This is more aligned with our conventions.
